Validate KDC support of S4U extensions
Per MS-SFU (see 1.5, 3.1.5.1.2 and 3.1.5.2.4):
Services can detect whether the KDC supports S4U by checking the cname
of the returned ticket. KDCs that do not support S4U ignore the
S4U2self and S4U2proxy data and return a service ticket with the cname
containing the name of the service that made the request ([RFC4120]
section 3.3.3). In service tickets from KDCs that support S4U, the
cname contains the name of the user.

Signed-off-by: Isaac Boukris <iboukris@gmail.com>