Enhance configure logic for auto-enabling otp
Since commit 3a0bf8eb0cd0fef82d5b41eafd7c74cb33fa1d2f, sqlite has been a
valid option for the HDB format (and thus can be the value of $db_type),
but we do not have a dbm compatibility wrapper for sqlite.  Similarly,
since commit a94dc50114e7cb32974b14015a18d4a9fef673c4, lmdb has been an
option, and again we do not haev a dbm compatibility wrapper for lmdb.
As such, the sqlite and lmdb db_types are incompatible with libotp,
since libotp has a hard requirement on the dbm family of APIs.  Alter the
logic in otp.m4 to reflect that all of the "unknown", "lmdb", and
"sqlite" db_types are incompatible with libotp (instead of just
"unknown"), and we must either error out (if otp was explicitly
requested) or disable otp (if using autodetection) if they are selected.

We also take this opportunity to note that since sqlite is available
in-tree, it will always be picked if it is in the preference list and no
external database type is available.  Thus, the only way for the
$db_type to be "unknown" after the rk_DB macro has run is if
--with-db-type-preference was explicitly provided by the user and sqlite
was excluded.  This leaves some of the logic within rk_DB as almost dead
code as well, so leave a comment to that effect to help the reader
understand the flow of logic.