gss: GSS_Indicate_mechs_by_attrs backwards compat
If a loadable mechanism does not support GSS_Inquire_attrs_for_mech() and no
attributes were marked as critical, include it unconditionally in the set of
returned mechanisms.

This is to avoid nasty surprises with mechanisms that do not support RFC 5587,
as future versions of Heimdal may more aggressively prune mechanisms from
negotiation based on attribute support. (The main deployed third-party
mechanisms are Moonshot and Samba NTLM; both of these do support
GSS_Inquire_attrs_for_mech(), so this is defensive programming only.)