krb5: krb5_config_parse_debug initialize output parameters
The caller of krb5_config_parse_debug() expects the output
parameters to be initialized even when the return code is non-zero.
This change initializes the output parameters in case the caller
did not.  Not all code paths assign values to the output parameters
which can result in unexpected termination of the process when
an uninitialized stack pointer is assumed to be valid.

Change-Id: I91859bb5a79ec08b8168e503e267070e2d9315a0