base: require platform/compiler support for atomics
If there are no platform or compiler atomic intrinsics, we currently fallback
to a single global mutex for all atomic operations. We anticipate that all
modern platforms support atomic operations, and if we don't currently have
support, it would be better to add them rather than falling back to this
unperformant workaround.

For the record, the current implementation supports C11, GCC, Solaris, AIX and
Windows atomics (used in that order of preference).